redirect and popup flow examples merged for next js
