+++ examples/nextjs-app/pages/index.tsx
@@ -1,20 +1,32 @@
-import Head from 'next/head'
-import styles from '../styles/Home.module.css'
+/* eslint-disable class-methods-use-this */
+import React from "react";
+import Link from "next/link";
 
-export default function Home() {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Torus </title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+interface IState {}
 
-      <main>
-        <div>
-          Example of torus-direct-web-sdk implementation in NextJS.
-        </div>
-      </main>
-    </div>
-  )
+interface IProps {}
+
+class HomePage extends React.PureComponent<IProps, IState> {
+  render() {
+    return (
+      <div
+        style={{
+          display: "flex",
+          flexDirection: "row",
+          justifyContent: "center",
+          alignItems: "center",
+          margin: 100,
+        }}
+      >
+        <Link href="/redirectMode/login">
+          <button>Login with Redirect Mode (Recommended)</button>
+        </Link>
+        <Link href="/popupMode/login">
+          <button>Login with Popup Mode</button>
+        </Link>
+      </div>
+    );
+  }
 }
+
+export default HomePage;
