+++ examples/nextjs-app/styles/Home.module.css
@@ -1,4 +1,4 @@
-.container {
+/* .container {
   min-height: 100vh;
   padding: 0 0.5rem;
   display: flex;
@@ -118,4 +118,4 @@
     width: 100%;
     flex-direction: column;
   }
-}
+} */
