+++ examples/nextjs-app/styles/globals.css
@@ -1,15 +1,26 @@
 body {
-  margin: 10px;
-  font-family: Avenir, Helvetica, Arial, sans-serif;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+}
+
+.App {
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
+  margin-top: 60px;
 }
-
-.console {
+#console {
   border: 1px solid black;
-  height: 80px;
+  height: 40px;
   padding: 2px;
   bottom: 10px;
   position: absolute;
@@ -17,13 +28,19 @@ body {
   width: calc(100% - 20px);
   border-radius: 5px;
 }
-.console::before {
-  content: "Console :";
+#console::before {
+  content: 'Console :';
   position: absolute;
   top: -20px;
   font-size: 12px;
 }
-.console > p {
+#console > p {
   margin: 0.5em;
-  word-wrap: break-word;
+}
+button {
+  height: 25px;
+  margin: 5px;
+  background: none;
+  border-radius: 5px;
+  cursor: pointer;
 }
